+++ packages/ethereum-contracts/contracts/agreements/gdav1/GeneralDistributionAgreementV1.sol
@@ -421,7 +421,9 @@ contract GeneralDistributionAgreementV1 is AgreementBase, TokenMonad, IGeneralDi
 
         newCtx = ctx;
 
-        if (_isPool(token, address(pool)) == false) {
+        if (_isPool(token, address(pool)) == false ||
+            // Note: we do not support multi-tokens pools
+            pool.superToken() != token) {
             revert GDA_ONLY_SUPER_TOKEN_POOL();
         }
 
@@ -485,7 +487,9 @@ contract GeneralDistributionAgreementV1 is AgreementBase, TokenMonad, IGeneralDi
         int96 requestedFlowRate,
         bytes calldata ctx
     ) external override returns (bytes memory newCtx) {
-        if (_isPool(token, address(pool)) == false) {
+        if (_isPool(token, address(pool)) == false ||
+            // Note: we do not support multi-tokens pools
+            pool.superToken() != token) {
             revert GDA_ONLY_SUPER_TOKEN_POOL();
         }
         if (requestedFlowRate < 0) {